RA-1266 Minor renaming change
