+++ src/Esfa.Vacancy.Register.UnitTests/SearchApprenticeship/Application/GivenAVacancySearchParametersMapper/AndFrameworkCodes.cs
@@ -3,9 +3,8 @@ using System.Linq;
 using Esfa.Vacancy.Register.Application.Queries.SearchApprenticeshipVacancies;
 using FluentAssertions;
 using NUnit.Framework;
-using NUnit.Framework.Internal;
 
-namespace Esfa.Vacancy.Register.UnitTests.SearchApprenticeship.Application.GivenVacancySearchParametersMapper
+namespace Esfa.Vacancy.Register.UnitTests.SearchApprenticeship.Application.GivenAVacancySearchParametersMapper
 {
     [TestFixture]
     public class AndFrameworkCodes
